New function create_input_panel
The new function create_input_panel includes the common code for
wAdvancedInputDialog and wInputDialog.

This function handles the panel, and return the text typed by the user.

The original functions continue doing the different code:

- wAdvancedInputDialog handles the history and change the pointers.
- wInputDialog only change the pointers.

I cannot change the pointer movement because the user can hit "Cancel"
in the panel and there is no way to know in wAdvancedInputDialog if
we should save the history, then wmaker crash.