app menu is destroyed not unmapped
This patch changes the UnMap menu with destroy menu. Now the menu
is destroyed when is not used (window is not selected). We don't
need refresh the menu or create it when the application is lanched.

wMenuUnmap not needed

This wMenuUnmap call is not needed. We can remove it. The unmap is
dode in the destroy.