New switch menu for rootmenu
This patch includes a new switch menu, different of switchmenu in
the rootmenu. Now we can create or destroy the switch menus independently.

The only difference with the wmaker original behaviour is that now we can
have two switch menus, one come from the root menu and other form OpenSwitchMenu
call.