Do not create sm and rm in startup
This patch remove the switchmenu and rootmenu creation in the startup.
Now, the pointers are empty and the menus are only created if the user
call them.

If the rootmenu was opened when the user restarts windowmaker, the menu
will continue opened. This behaviour is different in switchmenu, probably
we should use the same behaviour for both menus.