lvscr: appicon_map uses the vscr in creation
This patch start a serie of patches where the virtual screen used to map is the
virtual screen used when the object was created.

In the future, if we want change the virtual screen, we should include new functions
with unmap and remap in the new virtual screen.

The code is now more consistent and some function won't include the virtual screen,
because this variable is in the object already.