menu without entries remove X_ConfigureWindow warning
I discovered a warning in the patch:
    rcs: Remove set size values on core creation

The problem is when the menu has 0 items, the code makes a configure window
with "-1" as height argument. I tryed to remove the " -1" but is very complicated
and I got a lot of problems. So, I prefer do this hack to avoid this warning.

We should remove the "- 1" in the function

static int get_menu_height(WMenu *menu)
{
	return menu->entry_no * menu->entry_height - 1;
}

and do only:

static int get_menu_height(WMenu *menu)
{
        return menu->entry_no * menu->entry_height;
}

But this is another tale...

Warning info and the patch is this:

    rcs: Remove set size values on core creation
    The size values in the core struct are not used anymore, so we can remove them. The first
    step is remove from the creation. This patch does it.

    This patch creates a bug when the window list is empty and the user calls the switchmenu.

    /usr/local/bin/awmaker(catchXError(startup.c:119)): warning: internal X error: BadValue (integer parameter out of range for operation)
            Request code: 12 X_ConfigureWindow
            Request minor code: 0
            Resource ID: 0x0
            Error serial: 2585