BUG: crash when menu motion and empty menus
If the menu has no entries (switchmenu and no windows opened) and the menu is attached
to the rootmenu, if the user moves the mouse over the rootmenu, open the switchmenu as submenu
with motion (right click mouse always pushed), and then moves to the switchmenu title,
awmaker crash.

This patch checks if the menu has entries to avoid the crash.