fws: left_button_map uses static values
The left button is painted always in 0,0 in the titlebar. If we are using the
TS_NEW theme, se use these values. If we are not using TS_NEW (TS_OLD...) then
we use:

- For X: 0 + pad value (currently 3)
- For Y: title_height - button_height (button size). We split in 2 to center it.

So we can use these values in this function.

I am using the define TS_NORMAL_PAD for the padding.