fws: language_button_map uses static values
The language button is painted always at right of left button in the titlebar.
If we are using the TS_NEW theme, we use these values:

- For X:
  if not left button: 0
  if left button: btn_size (just at right of left button)
- For Y: 0 (the title bar starting point at Y, the top)

If we are not using TS_NEW (TS_OLD...) then we use an extra pad:

- For X:
  if not left button: 0 + pad value (currently 3)
  if left button: btn_size + pad value (currently 3)
- For Y: title_height - button_height (button size). We split in 2 to center it.

So we can use these values in this function.