fws: Move lines to group code
This patch only move lines.

First, move the line about the rigth button position to the top, because is
not used by other lines.

Then move the left and right button related code for the titlebar position
just before the right button position.

Then, the XKB_BUTTON_HINT related blocks are together, so we can join both
blocks and move the TS_NEW code to this block in the if-else.