fws: Calculate the titlebar width (step 1)
This patch calculates the titlebar width. If we are using theme TS_NORMAL
then the titlebar is painted in the framewin core, and we don't need calculate
it.

But if we are using the TS_NEW theme, the titlebar is painted using fwin->
titlebar, and then we need to move the titlebar next to the buttons. We need
set the starting point, and it depends on the number of buttons at left
(left button + language button). Also, we need set the total width, it
depends also en the buttons (left, language and rigth).