framewin: removed unused flag hide_language_button
The flag hide_language_button is set when the language_button_map flag is not
set and it is not set when the flag language_button_map is set.

Both flags are always used in the same places, together, so both flags have
the same information. We can remove one of them.