appicon menu is now destroyed
This patch destroyes the menu when exit the selection. I need include a lot
of changes:

1. If we destroy the menu and set the pointer to NULL the menu
will crash, because the entry 1 was previously wfree()ed and then
the wMenuDestroy will crash with double wfree() call in this menu

2. I add the argument bool to set the hidden or unhidden text in the entry 1,
and then we don't need use the wfree call in the if block.