Windows Menu in Root menu creation without X
The Windows Menu in Root menu is created but not mapped.
The mapping is done now in the addItem function, but not in the
menu creation.

The call to move menu is removed, because it crash awmaker because the
menu is not mapped yet, and we don't need move it (is not important).