Move code from OpenRootMenu to restore_rootmenu
This patch moves the lines of OpenRootMenu used in restore_rootmenu.
The patch removes the lines related to the keyboard (is False), updates
the values for x and y position.

Probably we can remove the values related to rebinding, but I am not sure
and this patch has a lot of changes.

Finally, I need to make public the functions used here, updating the
rootmenu.h file and removing some static strings.