WScreen includes virtual_screen
This patch creates two way pointer to link vscr to scr and scr to
vscr.

This is a very importan change in the code and the reason is because
every WScreen needs a virtual_screen, but we can have virtual_screens
without WScreen (are not painted).

vscr->screen_ptr could be null in a future, but
scr->vscr always will have a value.