def: set the window_place_origin
This patch removes the WScreen in the defaults to use it when the variable
is copied to the virtual_screen.

Perhaps, in a future we can hold the original value in the virtual_screen and
then modify to set in the WScreen.

I prefer to do the easier method in this patch.