default.c font behaviour changed
The behaviour of the Font is now different.

Now the function getFont saves in the wPreferences the font name.
Then, because we are not trying to get the XFont, we dont have X11
code here.

The functions to the specific fonts now reads the fontname from
the wPreferences and creates the font.