default.c xcolor behaviour changed
The behaviour of the XColor is now different.

Now the function getColor saves in the wPreferences the color name.
Then, because we are not trying to get the XColor, we dont have X11
code here.

The functions to the specific fonts now reads the colorname and the
default colorname from the wPreferences and creates the color.