default.c keybind behaviour changed
The behaviour of the keybinds is now different.

Now the function getCursor saves in the wPreferences the key name.
Then, because we are not trying to get the X calls, we dont have X11
code here.

The function setCursor select the right key bind stored in
wPreferences and create it.

Some things are not fine, I included some TODO to review it. awmaker
is working fine,... but could be better.