Remove the WDefaultEntry in update callback
This patch removes the WDefaultEntry in the update callback,
because is not used.

It is only used in some free calls, but I don't want to do it here,
because some functions are freeding the wPreferences field and other
not. The action should be the same in all cases. Is a TODO for the
future.