Added feature: automatic scaling of application icons
Based on the original patch from Tim Taenny <tim@taenny.de>

If the option "Enforce icon margin" is selected, application window
icons will be selected or scaled so that they only use 75% of the
available icon_size.
Even if the feature is not enabled, this change will scale down
large application icons to icon_size, so that icons can be used
that were rejected by the previous implementation of findBestIcon.
(Example: The Qt Creator icon never showed before, because it is
only provided in 128x128 resolution. Now it's visible.)

Signed-off-by: Rodolfo García Peñas (kix) <kix@kix.es>