Changed Legal and Info Panels
This patch has a lot of changes.

First, patches created by Tim Taenny <tim@taenny.de>

- - - -
Subject: Improved layout of the Legal Panel

Instead of relying on static pixel values for position and size of
the widgets, the legal panel now scales its widgets based on the
selected system font size.

Subject: Improved layout of the Info Panel

Instead of relying on static pixel values for position and size of
the widgets, the info panel now scales its widgets based on the
selected system font size.
- - - -

One of the patches (Legal) removes the WMGroove in the text and includes a
frame. I used the same idea for both panels (Info and Legal).

Then, I need to increase the margin size in the Info panel size. Then
changed the LEGAL_..._MARGIN variable to MARGIN because is used in both
panels (in the new frame).

Finally, increased a little bit the info panel height and updates the
Copyright to 2019.

I did all these changes together because the Tim patches shares the same code
in awmaker. In wmaker are different functions.

Small patch ;-)

Signed-off-by: Rodolfo García Peñas (kix) <kix@kix.es>