Split defaults convert and update. Unify the background logic
This patch joins the fucntions related to the background.
This is a very big patch, but introduces a lot of logic.

- First, we read all the defaults in optionList. When the defaults changes,
  then we set the refresh flag.
- Second, we check all the defualts in the optionList, if the flag is
  set, then refresh (call udpate) it.

The problem is the background, has a logic between the functions. The
functions uses a flag to check if the specific function was used or not
and then call the normal function to set the background. This is
critical, because we can launch the wmsetbg twice. I moved this logic to
only one function, that check both flags, and the make the job. Now we
don't need extra flags to do it.