Join update optionList functions
This patch joins the functions read_defaults_step1_update and
set_defaults_virtual_screen in the function set_defaults_virtual_screen.

They shared the same code, only the flag checking was different. This
patch adds the flag check in the set_defaults_virtual_screen function
and in the previous patches, the flag is initialized to one.