Enable workspace background refresh only in handler creation
This patch is related to the code deleted in the patch

Split defaults convert and update. Unify the background logic

This patch removed these lines in the function related to the convert
and update:

First block:

>--->--->---if (strcmp(entry->key, "WorkspaceBack") == 0 &&
->-->--->--- vscr->screen_ptr->flags.update_workspace_back &&
->-->--->--- vscr->screen_ptr->flags.backimage_helper_launched) {
->-->--->--->---needs_refresh |= default_update(vscr, entry, plvalue);
->-->--->---}

This code is used with the WorkspaceBack call. This call must be the
second in the optionList functions sort. The first is
WorkspaceSpecificBack. Then, if we are updating values, and the WorkspaceBack
was not updated, but the flag backimage_helper_launched is set, we need
call the WorkspaceBack->update function to refresh the default
background image. The code is used only if the
backimage_helper_launched is running.

The second block is used in the call to WorkspaceSpecificBack functions:

>--/*
>-- * If the WorkspaceSpecificBack data has been changed
>-- * so that the helper will be launched now, we must be
>-- * sure to send the default background texture config
>-- * to the helper.
>-- */
>--if (strcmp(entry->key, "WorkspaceSpecificBack") == 0 &&
>-- !vscr->screen_ptr->flags.backimage_helper_launched)
>-->---vscr->screen_ptr->flags.update_workspace_back = 1;

>--if (entry->refresh && entry->update) {
>-->---needs_refresh = (*entry->update) (vscr);
>-->---entry->refresh = 0;
>--}

This code is clear, we need set the flag if the helper is not launched
and we are updating the changed value of WorkspaceSpecificBack.

If the helper is running, then we don't need call WorkspaceBack.
This patch does exactly that, if we call WorkspaceSpecificBack->update
and the helper is not running, the cont flag is set to call
WorkspaceBack->update function, but not always (as previously).