New convert callback getClipMergedInDock
I included the new Callback getClipMergedInDock. This flag calls
getBool, as the previous convert callback, and then sets the flags like
the previous update callback.

Now is possible to know the merged_in_dock flag and the noclip flag in
the convert function.

Is important set the noclip flag first, because this flag is modified by
the getClipMergedInDock call.