New convert callback setIfDockPresent
I included the new Callback setIfDockPresent. This flag calls
getBool, as the previous convert callback, and then sets the flags like
the previous update callback.

Now is possible to know the no dock flag and the nodrawer flag in
the convert function.

Is important set the nodrawer flag first, because this flag is modified by
the setIfDockPresent call.