dock.c split
The file dock.c is huge. It has code for Clip, Dock and Drawers.

The code for the three elements is very similar, but I prefer split the
code in different files and hold the common code in dock-core file.

In this first change, I do not include different headers files, holding
only the dock.h file.