Use always the default icon
If the icon is not found, then we call get_rimage_icon_from_default_icon

This call was done by default (else block) and was used from the
wm_hints.

This call is not used (yet ;-) ) in the net_icon_image, because the
image was previously checked, the it doesn't matter.

This call was not used in the get_window_image_from_x11 call, and the
icon could be null. Now this problem is solved.