Do not paint icon twice
This patch removes these calls, because we are using them twice.

When we call wAppIcon_map(wapp->app_icon); we can find this call:

  map_icon_image(icon);

And then, these lines:

  WMAddNotificationObserver(icon_appearanceObserver, icon, WNIconAppearanceSettingsChanged, icon);
  WMAddNotificationObserver(icon_tileObserver, icon, WNIconTileSettingsChanged, icon);
  #ifdef USE_DOCK_XDND
  wXDNDMakeAwareness(aicon->icon->core->window);
  #endif
  AddToStackList(aicon->icon->vscr, aicon->icon->core);

These lines are related to callbacks DND and the stack.

In the call to map_icon_image, we can see:

void map_icon_image(WIcon *icon)
1 {
2 >---icon->file_image = get_rimage_from_file(icon->vscr, icon->file_name, wPreferences.icon_size);
3
4 >---/* Update the icon, because icon could be NULL */
5 >---wIconUpdate(icon);
6 >---wIconPaint(icon);
7 }

So the calls to wIconUpdate and wIconPaint are used twice. We can remove
them.